feat(beds): track individual beds and compute occupancy from them
Adds the bed table (room_id CASCADE, number, status, monthly_rent override) with UNIQUE(room_id, number). Migration 0004 applied. Beds are created with the room inside one transaction, so a room can never exist without them, and labelled A upward - a room reads 101-A, 101-B, 101-C.

Fixes the occupancy defect: both dashboard endpoints divided active tenants by ROOM count, so a 3-bed room with one tenant reported 100% occupied and an owner could not see two beds left to fill. Occupancy is now occupied beds over total beds, and both endpoints also return totalBeds and occupiedBeds. A maintenance bed stays in the denominator - it is still a bed the owner paid for and cannot currently sell. Negative control: restoring the old math fails 3 tests.

Capacity changes reconcile beds through a pure function: growing only appends, so an existing bed keeps its id, label, status and rent override and nobody is silently moved; shrinking removes the highest labels and is refused with a 409 naming the occupied beds, all or nothing. A gap left by a manually deleted bed is refilled. Room delete now refuses while any bed is occupied, since beds cascade with the room. New endpoints: GET /beds, GET /beds/vacant (excludes maintenance), GET /beds/:id, PATCH /beds/:id/status, PUT /beds/:id for the rent override, and GET /rooms/:roomId/beds.

Also fixes a latent string/number bug the occupancy test surfaced: count(*) and sum() return bigint/numeric, which pg hands back as strings, so totalRooms came out as '1'. Eleven aggregates across dashboard, admin, tenants, billing, payments and public now cast ::int - two of them fed a capacity comparison and one wrote paid_amount.

Adds backfill:beds for rooms predating the table, extends inspect:constraints with bedless-room and bed-count-vs-capacity checks, and on the frontend adds bed pills (filled occupied, outline vacant, amber maintenance) with a 44px touch target, a status dialog, and per-floor occupied/total counts. 13 unit tests on labelling and reconciliation, 16 API integration tests, 4 more on the structure totals.

+++ apps/api/src/routes/admin.ts
@@ -22,23 +22,23 @@ async function requireSuperAdmin(
 router.get("/overview", requireAuth, requireSuperAdmin, async (req: AuthenticatedRequest, res) => {
   try {
     const { userCount } = aggregate(
-      await db.select({ userCount: sql<number>`count(*)` }).from(user),
+      await db.select({ userCount: sql<number>`count(*)::int` }).from(user),
       { userCount: 0 },
     );
 
     const { ownerCount } = aggregate(
-      await db.select({ ownerCount: sql<number>`count(*)` }).from(ownerProfile),
+      await db.select({ ownerCount: sql<number>`count(*)::int` }).from(ownerProfile),
       { ownerCount: 0 },
     );
 
     const { propertyCount } = aggregate(
-      await db.select({ propertyCount: sql<number>`count(*)` }).from(property),
+      await db.select({ propertyCount: sql<number>`count(*)::int` }).from(property),
       { propertyCount: 0 },
     );
 
     const { tenantCount } = aggregate(
       await db
-        .select({ tenantCount: sql<number>`count(*)` })
+        .select({ tenantCount: sql<number>`count(*)::int` })
         .from(tenant)
         .where(eq(tenant.status, "active")),
       { tenantCount: 0 },
